+++ src/main/java/com/napzak/domain/member/api/service/MemberRegistrationService.java
@@ -1,8 +1,8 @@
-package com.napzak.domain.member.application;
+package com.napzak.domain.member.api.service;
 
 import com.napzak.domain.member.core.MemberEntity;
 import com.napzak.domain.member.core.Role;
-import com.napzak.domain.member.repository.MemberRepository;
+import com.napzak.domain.member.core.MemberRepository;
 import com.napzak.global.auth.client.dto.MemberSocialInfoResponse;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
