+++ src/main/java/com/napzak/domain/member/core/MemberRepository.java
@@ -1,12 +1,9 @@
-package com.napzak.domain.member.repository;
+package com.napzak.domain.member.core;
 
-import com.napzak.domain.member.core.MemberEntity;
-import com.napzak.domain.member.core.SocialType;
 import feign.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.lang.reflect.Member;
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
